Added multiple commands to handle the merge. Working on a query to copy the data, works with a single id but not multiple or something..
